studyflash-support/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
celery==5.4.0
redis==5.2.0
msal==1.31.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
websockets==13.1
openai==1.54.4
tenacity==9.0.0
python-dotenv==1.0.1
structlog==24.4.0
sentry-sdk==2.18.0
